Add example that separates label and input
This is quite an advanced use which will stretch the Design System to
the limit. HTML has long-supported placing labels far from the
corresponding input, even outside the form altogether. The risk here is
that error messages now will be placed by the input, which has a strong
likelihood of breaking layouts.

Still, I'm sure some people will find a use for this.

Fixes #477