Remove link to supplemental content
This change removes the aria-describedby link between labels/legends and
supplemental content, and the warnings that were added in #456.

It still allows arbitrary HTML content to be passed in, and it'll be
rendered in the same place - but importantly, it won't be read out by
screen readers. This can be a frustrating experience for screenr eader
users if there's multiple paragraphs read out between the label and the
input itself.

Fixes #445.